// src/services/types/useCase/home.ts
export type HomeApiResponse = {
  _id: string
  home: {
    section1: {
      image1: string
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      text1_en: string
      text1_fa: string
      button1_en: string
      button1_fa: string
      button2_en: string
      button2_fa: string
      text2_en: string
      text2_fa: string
    }
    section2: {
      title1_en: string
      title1_fa: string
      image1: string
      image2: string
      image3: string
      image4: string
      image5: string
      image6: string
    }
    section3: {
      image1: string
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      text1_en: string
      text1_fa: string
      button1_en: string
      button1_fa: string
      text2_en: string
      text2_fa: string
      text3_en: string
      text3_fa: string
      image2: string
    }
    section4: {
      image1: string
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      text1_en: string
      text1_fa: string
      button1_en: string
      button1_fa: string
      text2_en: string
      text2_fa: string
      text3_en: string
      text3_fa: string
      image2: string
    }
    section5: {
      image1: string
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      text1_en: string
      text1_fa: string
      button1_en: string
      button1_fa: string
      text2_en: string
      text2_fa: string
      text3_en: string
      text3_fa: string
      image2: string
    }
    section6: {
      image1: string
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      text1_en: string
      text1_fa: string
      button1_en: string
      button1_fa: string
      text2_en: string
      text2_fa: string
      text3_en: string
      text3_fa: string
      image2: string
    }
    section7: {
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      text3_en: string
      text3_fa: string
      button1_en: string
      button1_fa: string
    }
    section8: {
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      button1_en: string
      button1_fa: string
    }
  }
}